@@ -48,58 +49,58 @@
     "acorn": "8.15.0",
     "electron-settings": "4.0.4",
     "jquery": "3.7.1",
-    "mysql2": "3.15.0",
+    "mysql2": "3.15.3",
     "ngx-select-ex": "19.0.5",
     "ngx-toastr": "19.1.0",
     "ngx-ui-switch": "16.1.0",
-    "serialize-javascript": "6.0.2",
+    "serialize-javascript": "7.0.0",
     "sqlite3": "5.1.7",
     "tslib": "2.8.1"
   },
   "devDependencies": {
     "@angular-builders/custom-esbuild": "20.0.0",
-    "@angular-devkit/build-angular": "20.3.2",
-    "@angular-devkit/core": "20.3.2",
-    "@angular-devkit/schematics": "20.3.2",
-    "@angular-eslint/eslint-plugin": "20.3.0",
-    "@angular-eslint/eslint-plugin-template": "20.3.0",
-    "@angular-eslint/template-parser": "20.3.0",
-    "@angular/animations": "20.3.1",
-    "@angular/build": "20.3.2",
-    "@angular/common": "20.3.1",
-    "@angular/compiler": "20.3.1",
-    "@angular/compiler-cli": "20.3.1",
-    "@angular/core": "20.3.1",
-    "@angular/forms": "20.3.1",
-    "@angular/language-service": "20.3.1",
-    "@angular/platform-browser": "20.3.1",
-    "@angular/platform-browser-dynamic": "20.3.1",
-    "@angular/router": "20.3.1",
-    "@fortawesome/fontawesome-free": "7.0.1",
-    "@nx/angular": "21.5.3",
-    "@nx/eslint": "21.5.3",
-    "@nx/eslint-plugin": "21.5.3",
-    "@nx/js": "21.5.3",
-    "@nx/workspace": "21.5.3",
-    "@playwright/test": "1.55.0",
-    "@schematics/angular": "20.3.2",
+    "@angular-devkit/build-angular": "20.3.7",
+    "@angular-devkit/core": "20.3.7",
+    "@angular-devkit/schematics": "20.3.7",
+    "@angular-eslint/eslint-plugin": "20.5.0",
+    "@angular-eslint/eslint-plugin-template": "20.5.0",
+    "@angular-eslint/template-parser": "20.5.0",
+    "@angular/animations": "20.3.7",
+    "@angular/build": "20.3.7",
+    "@angular/common": "20.3.7",
+    "@angular/compiler": "20.3.7",
+    "@angular/compiler-cli": "20.3.7",
+    "@angular/core": "20.3.7",
+    "@angular/forms": "20.3.7",
+    "@angular/language-service": "20.3.7",
+    "@angular/platform-browser": "20.3.7",
+    "@angular/platform-browser-dynamic": "20.3.7",
+    "@angular/router": "20.3.7",
+    "@fortawesome/fontawesome-free": "7.1.0",
+    "@nx/angular": "21.6.6",
+    "@nx/eslint": "21.6.6",
+    "@nx/eslint-plugin": "21.6.6",
+    "@nx/js": "21.6.6",
+    "@nx/workspace": "21.6.6",
+    "@playwright/test": "1.56.0",
+    "@schematics/angular": "20.3.7",
     "@siemens/ngx-datatable": "22.4.1",
     "@swc-node/register": "1.11.1",
     "@swc/core": "1.13.5",
     "@swc/helpers": "0.5.17",
-    "@types/jasmine": "5.1.9",
+    "@types/jasmine": "5.1.12",
     "@types/jasminewd2": "2.0.13",
     "@types/jquery": "3.5.33",
     "@types/mysql": "2.15.27",
-    "@types/node": "22.18.6",
+    "@types/node": "24.10.0",
     "@types/sqlite3": "5.1.0",
     "@types/sqlstring": "2.3.2",
     "@typescript-eslint/eslint-plugin": "7.18.0",
     "@typescript-eslint/parser": "7.18.0",
     "bootstrap": "5.3.8",
     "conventional-changelog-cli": "4.1.0",
-    "cross-env": "10.0.0",
-    "electron": "37.5.1",
+    "cross-env": "10.1.0",
+    "electron": "37.7.1",
     "electron-builder": "26.0.12",
     "electron-reload": "1.5.0",
     "escape-string-regexp": "5.0.0",
