@@ -1,6 +1,8 @@
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { FormControl, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
+import { TranslateModule } from '@ngx-translate/core';
 import { SAI_TYPES, SmartScripts } from '@keira/shared/acore-world-model';
+import { WIKI_BASE_URL } from '@keira/shared/constants';
 import { CreatureSelectorBtnComponent, GameobjectSelectorBtnComponent } from '@keira/shared/selectors';
 
 import { SaiHandlerService } from '@keira/shared/sai-editor';
@@ -11,11 +13,12 @@ import { getEnumKeys, ModelForm } from '@keira/shared/utils';
   selector: 'keira-sai-search-entity',
   templateUrl: './sai-search-entity.component.html',
   styleUrls: ['./sai-search-entity.component.scss'],
-  imports: [FormsModule, ReactiveFormsModule, CreatureSelectorBtnComponent, GameobjectSelectorBtnComponent],
+  imports: [FormsModule, ReactiveFormsModule, TranslateModule, CreatureSelectorBtnComponent, GameobjectSelectorBtnComponent],
 })
 export class SaiSearchEntityComponent {
   private readonly handlerService = inject(SaiHandlerService);
 
+  readonly WIKI_BASE_URL = WIKI_BASE_URL;
   readonly SAI_SEARCH_TYPES = SAI_TYPES;
   readonly SAI_SEARCH_TYPES_KEYS = getEnumKeys(SAI_TYPES);
 
