@@ -1,15 +1,15 @@
-import { TestBed } from '@angular/core/testing';
 import { provideZonelessChangeDetection } from '@angular/core';
+import { TestBed } from '@angular/core/testing';
 import { provideNoopAnimations } from '@angular/platform-browser/animations';
 import { PageObject, Spied, TranslateTestingModule } from '@keira/shared/test-utils';
-import { ConnectionOptions, QueryError } from 'mysql2';
+import { QueryError } from 'mysql2';
 import { TooltipModule } from 'ngx-bootstrap/tooltip';
 import { of, throwError } from 'rxjs';
 import { instance, mock } from 'ts-mockito';
 
-import { ConnectionWindowComponent } from './connection-window.component';
+import { KeiraConnectionOptions, MysqlService } from '@keira/shared/db-layer';
 import { LoginConfigService } from '@keira/shared/login-config';
-import { MysqlService } from '@keira/shared/db-layer';
+import { ConnectionWindowComponent } from './connection-window.component';
 
 class ConnectionWindowComponentPage extends PageObject<ConnectionWindowComponent> {
   get hostInput(): HTMLInputElement {
@@ -36,6 +36,9 @@ class ConnectionWindowComponentPage extends PageObject<ConnectionWindowComponent
   get savePasswordInput(): HTMLInputElement {
     return this.query<HTMLInputElement>('#save-password');
   }
+  get sslEnabledInput(): HTMLInputElement {
+    return this.query<HTMLInputElement>('#ssl-enabled');
+  }
   get rememberMeInput(): HTMLInputElement {
     return this.query<HTMLInputElement>('#remember-me');
   }
@@ -51,13 +54,14 @@ class ConnectionWindowComponentPage extends PageObject<ConnectionWindowComponent
 }
 
 describe('ConnectionWindowComponent', () => {
-  const mockConfigsWithPass: Partial<ConnectionOptions>[] = [
+  const mockConfigsWithPass: Partial<KeiraConnectionOptions>[] = [
     {
       host: 'localhost',
       port: 3306,
       user: 'Shin',
       password: 'shin123',
       database: 'shin_world',
+      sslEnabled: false,
     },
     {
       host: '127.0.0.1',
@@ -65,15 +69,17 @@ describe('ConnectionWindowComponent', () => {
       user: 'Helias',
       password: 'helias123',
       database: 'helias_world',
+      sslEnabled: false,
     },
   ];
-  const mockConfigsNoPass: Partial<ConnectionOptions>[] = [
+  const mockConfigsNoPass: Partial<KeiraConnectionOptions>[] = [
     {
       host: 'localhost',
       port: 3306,
       user: 'Shin',
       password: 'shin123',
       database: 'shin_world',
+      sslEnabled: false,
     },
     {
       host: '127.0.0.1',
@@ -81,6 +87,7 @@ describe('ConnectionWindowComponent', () => {
       user: 'Helias',
       password: '',
       database: 'helias_world',
+      sslEnabled: false,
     },
   ];
 
@@ -125,6 +132,7 @@ describe('ConnectionWindowComponent', () => {
     const { page, component, connectSpy } = setup();
     component.error = { code: 'some previous error', errno: 1234 } as QueryError;
 
+    page.sslEnabledInput.click();
     page.clickElement(page.connectBtn);
 
     expect(connectSpy).toHaveBeenCalledTimes(1);
@@ -134,6 +142,8 @@ describe('ConnectionWindowComponent', () => {
       user: 'root',
       password: 'root',
       database: 'acore_world',
+      sslEnabled: true,
+      ssl: { rejectUnauthorized: false },
     });
     expect(component.error).toBeUndefined();
     expect(page.errorElement.innerHTML).not.toContain('error-box');
@@ -154,6 +164,7 @@ describe('ConnectionWindowComponent', () => {
       user: 'Helias',
       password: 'helias123',
       database: 'helias_world',
+      sslEnabled: false,
     });
 
     expect(component.error).toBeUndefined();
@@ -177,7 +188,7 @@ describe('ConnectionWindowComponent', () => {
     page.clickElement(page.connectBtn);
 
     expect(connectSpy).toHaveBeenCalledTimes(1);
-    expect(connectSpy).toHaveBeenCalledWith({ host, port, user, password, database });
+    expect(connectSpy).toHaveBeenCalledWith({ host, port, user, password, database, sslEnabled: false });
     expect(component.error).toBeUndefined();
     expect(page.errorElement.innerHTML).not.toContain('error-box');
   });
@@ -254,6 +265,7 @@ describe('ConnectionWindowComponent', () => {
         user: 'Helias',
         password,
         database: 'helias_world',
+        sslEnabled: false,
       });
       expect(connectSpy).toHaveBeenCalledTimes(1);
       expect(connectSpy).toHaveBeenCalledWith({
@@ -262,6 +274,7 @@ describe('ConnectionWindowComponent', () => {
         user: 'Helias',
         password,
         database: 'helias_world',
+        sslEnabled: false,
       });
     });
 
@@ -281,6 +294,7 @@ describe('ConnectionWindowComponent', () => {
         user: 'Helias',
         password: '',
         database: 'helias_world',
+        sslEnabled: false,
       });
       expect(connectSpy).toHaveBeenCalledTimes(1);
       expect(connectSpy).toHaveBeenCalledWith({
@@ -289,6 +303,7 @@ describe('ConnectionWindowComponent', () => {
         user: 'Helias',
         password: 'helias123',
         database: 'helias_world',
+        sslEnabled: false,
       });
     });
   });
@@ -340,7 +355,7 @@ describe('ConnectionWindowComponent', () => {
       //  but we can't because of: https://github.com/valor-software/ngx-bootstrap/issues/4282
     });
 
-    it('clear all should correctly work', async () => {
+    it('clear all should correctly work', () => {
       const { page, component, loginConfigService } = setup();
 
       // TODO: ideally we should click the dropdown's item instead of calling the method
@@ -356,7 +371,7 @@ describe('ConnectionWindowComponent', () => {
       expect(page.databaseInput.value).toBe('');
     });
 
-    it('loadConfig should correctly work', async () => {
+    it('loadConfig should correctly work', () => {
       const { page, component } = setup();
 
       // TODO: ideally we should click the dropdown's item instead of calling the method
