@@ -1,17 +1,18 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, inject, OnInit } from '@angular/core';
 import { FormControl, FormGroup, FormsModule, ReactiveFormsModule } from '@angular/forms';
 
-import { ConnectionOptions, QueryError } from 'mysql2';
+import { QueryError } from 'mysql2';
 // eslint-disable-next-line @nx/enforce-module-boundaries
 import packageInfo from '../../../../package.json';
 
 import { TranslateModule } from '@ngx-translate/core';
 import { BsDropdownModule } from 'ngx-bootstrap/dropdown';
+import { TooltipModule } from 'ngx-bootstrap/tooltip';
 
+import { QueryErrorComponent } from '@keira/shared/base-editor-components';
+import { KeiraConnectionOptions, MysqlService } from '@keira/shared/db-layer';
 import { LoginConfigService } from '@keira/shared/login-config';
 import { SwitchLanguageComponent } from '@keira/shared/switch-language';
-import { MysqlService } from '@keira/shared/db-layer';
-import { QueryErrorComponent } from '@keira/shared/base-editor-components';
 import { ModelForm, SubscriptionHandler } from '@keira/shared/utils';
 
 @Component({
@@ -19,7 +20,15 @@ import { ModelForm, SubscriptionHandler } from '@keira/shared/utils';
   selector: 'keira-connection-window',
   templateUrl: './connection-window.component.html',
   styleUrls: ['./connection-window.component.scss'],
-  imports: [FormsModule, ReactiveFormsModule, BsDropdownModule, TranslateModule, QueryErrorComponent, SwitchLanguageComponent],
+  imports: [
+    FormsModule,
+    ReactiveFormsModule,
+    BsDropdownModule,
+    TranslateModule,
+    TooltipModule,
+    QueryErrorComponent,
+    SwitchLanguageComponent,
+  ],
 })
 export class ConnectionWindowComponent extends SubscriptionHandler implements OnInit {
   private readonly mysqlService = inject(MysqlService);
@@ -29,8 +38,8 @@ export class ConnectionWindowComponent extends SubscriptionHandler implements On
   private readonly IMAGES_COUNT = 11;
   readonly RANDOM_IMAGE = Math.floor(Math.random() * this.IMAGES_COUNT) + 1;
   readonly KEIRA_VERSION = packageInfo.version;
-  configs: Partial<ConnectionOptions>[] | undefined;
-  form!: FormGroup<ModelForm<Partial<ConnectionOptions>>>;
+  configs: Partial<KeiraConnectionOptions>[] | undefined;
+  form!: FormGroup<ModelForm<Partial<KeiraConnectionOptions>>>;
   error: QueryError | undefined;
   savePassword = true;
   rememberMe = true;
@@ -40,19 +49,21 @@ export class ConnectionWindowComponent extends SubscriptionHandler implements On
   }
 
   ngOnInit(): void {
-    this.form = new FormGroup<ModelForm<Partial<ConnectionOptions>>>({
+    this.form = new FormGroup<ModelForm<Partial<KeiraConnectionOptions>>>({
       host: new FormControl<string>('127.0.0.1') as FormControl<string>,
       port: new FormControl<number>(3306) as FormControl<number>,
       user: new FormControl<string>('root') as FormControl<string>,
       password: new FormControl<string>('root') as FormControl<string>,
       database: new FormControl<string>('acore_world') as FormControl<string>,
+      sslEnabled: new FormControl<boolean>(false) as FormControl<boolean>,
     });
 
     this.configs = this.loginConfigService.getConfigs();
 
     if (this.configs?.length > 0) {
       // get last saved config
-      this.form.setValue(this.configs[this.configs.length - 1]);
+      const lastConfig = this.configs[this.configs.length - 1];
+      this.form.setValue(lastConfig);
 
       if (!this.form.getRawValue().password) {
         this.savePassword = false;
@@ -64,7 +75,7 @@ export class ConnectionWindowComponent extends SubscriptionHandler implements On
     }
   }
 
-  loadConfig(config: Partial<ConnectionOptions>): void {
+  loadConfig(config: Partial<KeiraConnectionOptions>): void {
     this.form.setValue(config);
   }
 
@@ -75,13 +86,23 @@ export class ConnectionWindowComponent extends SubscriptionHandler implements On
   }
 
   onConnect(): void {
+    const connectionConfig = this.form.getRawValue();
+
+    // Add SSL configuration if SSL is enabled
+    if (this.form.getRawValue().sslEnabled) {
+      connectionConfig.ssl = {
+        rejectUnauthorized: false, // Allow self-signed certificates
+      };
+    }
+
     this.subscriptions.push(
-      this.mysqlService.connect(this.form.getRawValue()).subscribe({
+      this.mysqlService.connect(connectionConfig).subscribe({
         next: () => {
           const newConfig = this.form.getRawValue();
           if (!this.savePassword) {
             newConfig.password = '';
           }
+
           this.loginConfigService.saveRememberPreference(this.rememberMe);
           this.loginConfigService.saveNewConfig(newConfig);
           this.error = undefined;
