@@ -1,16 +1,16 @@
 import { Option } from '@keira/shared/constants';
 
 export const CREATURE_CLASS: Option[] = [
-  { value: 0, name: 'NONE' },
-  { value: 1, name: 'WARRIOR' },
-  { value: 2, name: 'PALADIN' },
-  { value: 3, name: 'HUNTER' },
-  { value: 4, name: 'ROGUE' },
-  { value: 5, name: 'PRIEST' },
-  { value: 6, name: 'DEATH_KNIGHT' },
-  { value: 7, name: 'SHAMAN' },
-  { value: 8, name: 'MAGE' },
-  { value: 9, name: 'WARLOCK' },
-  { value: 10, name: 'UNK' },
-  { value: 11, name: 'DRUID' },
+  { value: 0, name: 'NONE', icon: '' },
+  { value: 1, name: 'WARRIOR', icon: 'class/1.gif' },
+  { value: 2, name: 'PALADIN', icon: 'class/2.gif' },
+  { value: 3, name: 'HUNTER', icon: 'class/3.gif' },
+  { value: 4, name: 'ROGUE', icon: 'class/4.gif' },
+  { value: 5, name: 'PRIEST', icon: 'class/5.gif' },
+  { value: 6, name: 'DEATH_KNIGHT', icon: 'class/6.gif' },
+  { value: 7, name: 'SHAMAN', icon: 'class/7.gif' },
+  { value: 8, name: 'MAGE', icon: 'class/8.gif' },
+  { value: 9, name: 'WARLOCK', icon: 'class/9.gif' },
+  { value: 10, name: 'UNK', icon: '' },
+  { value: 11, name: 'DRUID', icon: 'class/11.gif' },
 ];
