@@ -1,27 +1,27 @@
 import { Option } from '@keira/shared/constants';
 
 export const CREATURE_RACE: Option[] = [
-  { value: 0, name: 'NONE' },
-  { value: 1, name: 'HUMAN' },
-  { value: 2, name: 'ORC' },
-  { value: 3, name: 'DWARF' },
-  { value: 4, name: 'NIGHTELF' },
-  { value: 5, name: 'UNDEAD_PLAYER' },
-  { value: 6, name: 'TAUREN' },
-  { value: 7, name: 'GNOME' },
-  { value: 8, name: 'TROLL' },
-  { value: 9, name: 'GOBLIN (NOT USED)' },
-  { value: 10, name: 'BLOODELF' },
-  { value: 11, name: 'DRAENEI' },
-  { value: 12, name: 'FEL_ORC (NOT USED)' },
-  { value: 13, name: 'NAGA (NOT USED)' },
-  { value: 14, name: 'BROKEN (NOT USED)' },
-  { value: 15, name: 'SKELETON (NOT USED)' },
-  { value: 16, name: 'VRYKUL (NOT USED)' },
-  { value: 17, name: 'TUSKARR (NOT USED)' },
-  { value: 18, name: 'FOREST_TROLL (NOT USED)' },
-  { value: 19, name: 'TAUNKA (NOT USED)' },
-  { value: 20, name: 'NORTHREND_SKELETON (NOT USED)' },
+  { value: 0, name: 'NONE', icon: '' },
+  { value: 1, name: 'HUMAN', icon: 'race/1-0.gif' },
+  { value: 2, name: 'ORC', icon: 'race/2-0.gif' },
+  { value: 3, name: 'DWARF', icon: 'race/3-0.gif' },
+  { value: 4, name: 'NIGHTELF', icon: 'race/4-0.gif' },
+  { value: 5, name: 'UNDEAD_PLAYER', icon: 'race/5-0.gif' },
+  { value: 6, name: 'TAUREN', icon: 'race/6-0.gif' },
+  { value: 7, name: 'GNOME', icon: 'race/7-0.gif' },
+  { value: 8, name: 'TROLL', icon: 'race/8-0.gif' },
+  { value: 9, name: 'GOBLIN (NOT USED)', icon: 'race/9-0.gif' },
+  { value: 10, name: 'BLOODELF', icon: 'race/10-0.gif' },
+  { value: 11, name: 'DRAENEI', icon: 'race/11-0.gif' },
+  { value: 12, name: 'FEL_ORC (NOT USED)', icon: '' },
+  { value: 13, name: 'NAGA (NOT USED)', icon: '' },
+  { value: 14, name: 'BROKEN (NOT USED)', icon: '' },
+  { value: 15, name: 'SKELETON (NOT USED)', icon: '' },
+  { value: 16, name: 'VRYKUL (NOT USED)', icon: '' },
+  { value: 17, name: 'TUSKARR (NOT USED)', icon: '' },
+  { value: 18, name: 'FOREST_TROLL (NOT USED)', icon: '' },
+  { value: 19, name: 'TAUNKA (NOT USED)', icon: '' },
+  { value: 20, name: 'NORTHREND_SKELETON (NOT USED)', icon: '' },
 ];
 
 export const CREATURE_RACE_OPTION_ICON: Option[] = [
@@ -35,4 +35,5 @@ export const CREATURE_RACE_OPTION_ICON: Option[] = [
   { value: 8, name: 'TROLL', icon: 'race/8-0.gif' },
   { value: 9, name: 'GOBLIN', icon: 'race/9-0.gif' },
   { value: 10, name: 'BLOODELF', icon: 'race/10-0.gif' },
+  { value: 11, name: 'DRAENEI', icon: 'race/11-0.gif' },
 ];
