@@ -1,10 +1,10 @@
 import { Option } from '@keira/shared/constants';
 
 export const CREATURE_TEMPLATE_RESISTANCE_SCHOOL: Option[] = [
-  { value: 1, name: 'SPELL_SCHOOL_HOLY' },
-  { value: 2, name: 'SPELL_SCHOOL_FIRE' },
-  { value: 3, name: 'SPELL_SCHOOL_NATURE' },
-  { value: 4, name: 'SPELL_SCHOOL_FROST' },
-  { value: 5, name: 'SPELL_SCHOOL_SHADOW' },
-  { value: 6, name: 'SPELL_SCHOOL_ARCANE' },
+  { value: 1, name: 'SPELL_SCHOOL_HOLY', icon: 'dmgschool/holy.png' },
+  { value: 2, name: 'SPELL_SCHOOL_FIRE', icon: 'dmgschool/fire.png' },
+  { value: 3, name: 'SPELL_SCHOOL_NATURE', icon: 'dmgschool/nature.png' },
+  { value: 4, name: 'SPELL_SCHOOL_FROST', icon: 'dmgschool/frost.png' },
+  { value: 5, name: 'SPELL_SCHOOL_SHADOW', icon: 'dmgschool/shadow.png' },
+  { value: 6, name: 'SPELL_SCHOOL_ARCANE', icon: 'dmgschool/arcane.png' },
 ];
