@@ -1,11 +1,11 @@
 import { Option } from '@keira/shared/constants';
 
 export const DAMAGE_TYPE: Option[] = [
-  { value: 0, name: 'NORMAL' },
-  { value: 1, name: 'HOLY' },
-  { value: 2, name: 'FIRE' },
-  { value: 3, name: 'NATURE' },
-  { value: 4, name: 'FROST' },
-  { value: 5, name: 'SHADOW' },
-  { value: 6, name: 'ARCANE' },
+  { value: 0, name: 'NORMAL', icon: 'dmgschool/normal.png' },
+  { value: 1, name: 'HOLY', icon: 'dmgschool/holy.png' },
+  { value: 2, name: 'FIRE', icon: 'dmgschool/fire.png' },
+  { value: 3, name: 'NATURE', icon: 'dmgschool/nature.png' },
+  { value: 4, name: 'FROST', icon: 'dmgschool/frost.png' },
+  { value: 5, name: 'SHADOW', icon: 'dmgschool/shadow.png' },
+  { value: 6, name: 'ARCANE', icon: 'dmgschool/arcane.png' },
 ];
