@@ -1,12 +1,12 @@
 import { Option } from '@keira/shared/constants';
 
 export const FACTION_RANK: Option[] = [
-  { value: 0, name: 'Hated' },
-  { value: 1, name: 'Hostile' },
-  { value: 2, name: 'Unfriendly' },
-  { value: 3, name: 'Neutral' },
-  { value: 4, name: 'Friendly' },
-  { value: 5, name: 'Honored' },
-  { value: 6, name: 'Revered' },
-  { value: 7, name: 'Exalted' },
+  { value: 0, name: 'Hated', icon: 'reputation/hated.png' },
+  { value: 1, name: 'Hostile', icon: 'reputation/hostile.png' },
+  { value: 2, name: 'Unfriendly', icon: 'reputation/unfriendly.png' },
+  { value: 3, name: 'Neutral', icon: 'reputation/neutral.png' },
+  { value: 4, name: 'Friendly', icon: 'reputation/friendly.png' },
+  { value: 5, name: 'Honored', icon: 'reputation/honored.png' },
+  { value: 6, name: 'Revered', icon: 'reputation/revered.png' },
+  { value: 7, name: 'Exalted', icon: 'reputation/exalted.png' },
 ];
