@@ -1,10 +1,10 @@
 import { Option } from '@keira/shared/constants';
 
 export const GAMEOBJECT_ICON: Option[] = [
-  { value: '', name: 'None' },
-  { value: 'Taxi', name: 'Taxi' },
-  { value: 'Talk', name: 'Talk' },
-  { value: 'Attack', name: 'Attack' },
-  { value: 'Directions', name: 'Directions' },
-  { value: 'Quest', name: 'Quest' },
+  { value: '', name: 'None', icon: '' },
+  { value: 'Taxi', name: 'Taxi', icon: 'gossip/taxi.png' },
+  { value: 'Talk', name: 'Talk', icon: 'gossip/talk.png' },
+  { value: 'Attack', name: 'Attack', icon: 'creature-icons/Attack.png' },
+  { value: 'Directions', name: 'Directions', icon: 'creature-icons/Directions.png' },
+  { value: 'Quest', name: 'Quest', icon: 'creature-icons/Quest.png' },
 ];
