@@ -1,14 +1,13 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, inject, Input, input } from '@angular/core';
 import { ICON_SKILLS } from '@keira/shared/constants';
-import { IconService } from './icon.service';
 import { SubscriptionHandler } from '@keira/shared/utils';
+import { IconService } from './icon.service';
 
 @Component({
   selector: 'keira-icon',
   templateUrl: './icon.component.html',
   styleUrls: ['./icon.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
-  standalone: true,
 })
 export class IconComponent extends SubscriptionHandler {
   private readonly service = inject(IconService);
