@@ -1,5 +1,6 @@
-export * from './sqlite.service';
+export * from './mysql.model';
 export * from './mysql.service';
-export * from './query/mysql-query.service';
 export * from './query/base-query.service';
+export * from './query/mysql-query.service';
 export * from './query/sqlite-query.service';
+export * from './sqlite.service';
