@@ -54,13 +54,13 @@ describe('MysqlQueryService', () => {
   });
 
   it('query() should call mysqlService.dbQuery() and output query and results if debug mode is enabled', () => {
-    const logSpy = spyOn(console, 'log');
+    const infoSpy = spyOn(console, 'info');
     configService.debugMode.set(true);
     const querySpy = spyOn(TestBed.inject(MysqlService), 'dbQuery').and.returnValue(of({ id: 'mock value' } as TableRow));
     const myQuery = 'SELECT azerothcore FROM projects;';
 
     service.query(myQuery).subscribe(() => {
-      expect(logSpy).toHaveBeenCalledTimes(2);
+      expect(infoSpy).toHaveBeenCalledTimes(2);
     });
 
     expect(querySpy).toHaveBeenCalledWith(myQuery, undefined);
@@ -800,6 +800,7 @@ describe('MysqlQueryService', () => {
       { name: 'getText0ById', query: `SELECT text0_0 AS v FROM npc_text WHERE ID = ${id}` },
       { name: 'getText1ById', query: `SELECT text0_1 AS v FROM npc_text WHERE ID = ${id}` },
       { name: 'getCreatureDisplayIdById', query: `SELECT CreatureDisplayID AS v FROM creature_template_model WHERE CreatureID=${id}` },
+      { name: 'getGameobjectDisplayIdById', query: `SELECT displayId AS v FROM gameobject_template WHERE entry=${id}` },
     ];
 
     for (const test of cases) {
